+++ src/components/icon/icon.tsx
@@ -3,6 +3,7 @@ import styled from 'styled-components';
 
 import CheckmarkCircle from '../../assets/icons/checkmarkCircle.svg';
 import Checkmark from '../../assets/icons/checkmarkSelected.svg';
+import CheckmarkBox from '../../assets/icons/checkmarkBox.svg';
 import ChevronDown from '../../assets/icons/chevronDown.svg';
 import ChevronLeft from '../../assets/icons/chevronLeft.svg';
 import ChevronRight from '../../assets/icons/chevronRight.svg';
@@ -27,6 +28,7 @@ const defaultColor = greys.shade50;
 const importedIcons = {
   CheckmarkCircle,
   CheckmarkSelected: Checkmark,
+  CheckmarkBox,
   ChevronDown,
   ChevronLeft,
   ChevronRight,
@@ -55,6 +57,8 @@ interface IconProps {
   color?: string;
   /** Controls the width and height of the icon. */
   size?: number;
+  /** If set, no color for the icon will be set. This enables inheriting the parents color. */
+  shouldDisableColor?: boolean;
 }
 
 /*
@@ -75,7 +79,7 @@ const StyledIconDiv = styled.div<StyledIconDivProps>`
  */
 
 export const Icon: FC<IconProps> = props => {
-  const {name, size = defaultSize, color = defaultColor} = props;
+  const {name, size = defaultSize, color = defaultColor, shouldDisableColor} = props;
 
   // Pull the icon from the list of available icons.
   // The svg file is converted to a function that is called.
@@ -88,7 +92,7 @@ export const Icon: FC<IconProps> = props => {
       {icon({
         width: size,
         height: size,
-        color,
+        color: shouldDisableColor ? '' : color,
         preserveAspectRatio: "none"
       })}
     </StyledIconDiv>