+++ src/index.ts
@@ -16,6 +16,13 @@ export {VisualSizesEnum, fonts} from './helpers/fontHelpers';
 
 export {Avatar} from './components/avatar/avatar';
 
+export {Button, ButtonTypes} from './components/button/button';
+export {ButtonContent} from './components/button/buttonContent';
+export {ButtonContentIcon} from './components/button/buttonContentIcon';
+export {ButtonGroup} from './components/button/buttonGroup';
+
 export {Checkbox} from './components/checkbox/checkbox';
 
 export {Icon, IconName} from './components/icon/icon';
+
+export {Skeleton} from './components/skeleton/skeleton';